refactor(ai-service): Step 2 — move feature catalogs to app/domain/
Domain types are leaves: no imports from app/services/. This step
moves the column-name enums and config dataclasses out of two
service files so segment_models.py can later land in app/domain/
without violating the layering rule.

Moved:
  app/services/imitate_expert_learning_service.py
    ├─ ExpertFeatureCatalog            ─┐
    ├─ SegmentImprovementConfig         │
    ├─ SegmentImprovementSummary        ├─►  app/domain/expert_features.py
    └─ ExpertModelTrainingConfig       ─┘

  app/services/tire_grip_analysis_service.py
    ├─ TireGripFeatureCatalog          ─┬─►  app/domain/tire_grip_features.py
    └─ SlipEnvelopeConfig              ─┘

Service files re-import these symbols at the top so existing internal
references and the public surface keep working unchanged. The helper
`_format_debug_message` (service-internal logging) and the constant
`EPS` (used inside the slip calculation) stay in the service files —
they have no callers outside their host file.

Importers updated to point at app.domain directly:
  • app/services/transformer_model.py
  • app/services/telemetry_segment_visualizer.py
  • app/models/segment_models.py  (resolves a domain→services back-edge
                                   flagged on Page 6 of the diagram)

Validated:
  • python -c "from app.domain.expert_features import …"   identity ✓
  • python -c "from app.domain.tire_grip_features import …" identity ✓
  • re-imports through services/ return the SAME class objects
  • app.main imports cleanly inside the container
  • /health returns 200 after uvicorn --reload picked up the change
