+++ acla_ai_service/app/services/tire_grip_analysis_service.py
@@ -24,33 +24,12 @@ import pandas as pd
 # Suppress warnings
 warnings.filterwarnings('ignore', category=UserWarning)
 
-class TireGripFeatureCatalog:
-    """Authoritative feature names emitted by TireGripAnalysisService."""
-
-    class ContextFeature(str, Enum):
-        DRIVER_PUSH_TO_LIMIT = 'driver_push_to_limit'
-
-    CONTEXT_FEATURES: List[str] = [f.value for f in ContextFeature]
-
+# Domain types moved to app/domain/ in refactor/hexagonal-v1 Step 2.
+# Re-imported here so internal references and external callers
+# (transformer_model, segment_models, telemetry_segment_visualizer) keep working.
+from app.domain.tire_grip_features import SlipEnvelopeConfig, TireGripFeatureCatalog
 
 EPS = 1e-6
-
-
-@dataclass
-class SlipEnvelopeConfig:
-    """Hyperparameters governing the slip-envelope calculation."""
-
-    front_slip_limit: float = 0.105  # ~6.0 degrees in radians
-    rear_slip_limit: float = 0.122   # ~7.0 degrees in radians
-    front_longitudinal_slip_limit: float = 0.1  # Dimensionless slip ratio
-    rear_longitudinal_slip_limit: float = 0.1
-    
-    # Weights for the combined metric
-    slip_angle_weight: float = 1.0
-    slip_ratio_weight: float = 1.0
-
-    def combined_limit(self) -> float:
-        return max(self.front_slip_limit, self.rear_slip_limit)
 
 
 class TireGripAnalysisService:
