+++ acla_ai_service/app/services/transformer_model.py
@@ -18,10 +18,8 @@ from sklearn.preprocessing import StandardScaler, MinMaxScaler
 from sklearn.metrics import mean_squared_error, mean_absolute_error
 
 # Import contextual feature catalogs for quality weighting
-from .tire_grip_analysis_service import TireGripFeatureCatalog
-from .imitate_expert_learning_service import (
-    ExpertFeatureCatalog,
-)
+from app.domain.tire_grip_features import TireGripFeatureCatalog
+from app.domain.expert_features import ExpertFeatureCatalog
 from ..models.telemetry_models import TelemetryFeatures, _safe_float
 
 # Force unbuffered output for real-time print statements
